Mostrar mensaje de confirmacion y espera.

+++ src/modules/daybook/views/EntryView.vue
@@ -35,6 +35,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '@/modules/daybook/helpers/getDayMonthYear'
 
@@ -85,6 +86,13 @@ export default {
             this.entry = entry
         },
         async saveEntry(){
+            new Swal({
+                title:'Espere por favor',
+                allowOutsideClick:false
+            })
+
+            Swal.showLoading()
+
             //console.log('Guardando entrada')
             if( this.entry.id ){
                 this.updateEntry( this.entry )
@@ -95,11 +103,30 @@ export default {
                 const id = await this.createEntry( this.entry )
                 this.$router.push({ name:'entry', params:{ id }})
             }
+
+            Swal.fire('Guardado', 'Entrada registrada con exito', 'success')
         },
         async onToDelete(){
-            //console.log('Borrando', this.entry)
-            await this.deleteEntry(this.entry.id)
-            this.$router.push({ name:'no-entry' })
+
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Esta seguro?',
+                text: 'una vez borrado, no se podra recuperar.',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro.'
+            })
+
+            if( isConfirmed ){
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                //console.log('Borrando', this.entry)
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({ name:'no-entry' })
+
+                Swal.fire('Eliminado', '', 'success')
+            }
         }
     },
     created(){
